requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
certifi==2024.8.30
charset-normalizer==2.0.12
environs==8.0.0
frozenlist==1.4.1
greenlet==3.1.0
idna==3.8
magic-filter==1.0.12
marshmallow==3.22.0
multidict==6.1.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.2
requests==2.26.0
SQLAlchemy==1.4.54
typing_extensions==4.12.2
urllib3==1.26.20
yarl==1.11.1
